move trajctory follow command file to commands folder
