+++ src/main/java/frc/robot/subsystems/drive/Drive.java
@@ -64,6 +64,7 @@ import frc.lib.util.PID;
 import frc.robot.Constants;
 import frc.robot.Constants.Mode;
 import frc.robot.RobotState;
+import frc.robot.commands.ResilientTrajectoryFollower;
 
 import org.littletonrobotics.junction.AutoLogOutput;
 import org.littletonrobotics.junction.Logger;
